Revert "Added images to database, and first view of details page"
This reverts commit 4d555648fade54af338bac019bfa1d855e463aeb.
