--- Web_Api_Amazon/wwwroot/css/product.css
@@ -1,444 +1,0 @@
-﻿
-:root {
-    --accent: #c7f04d;
-    --accent-dark: #8bbf20;
-    --ink: #111827;
-    --muted: #6b7280;
-    --border: #e5e7eb;
-    --card: #ffffff;
-    --panel: #f7f8fb;
-    --soft: #f1f5f9;
-}
-
-.product-page {
-    display: flex;
-    flex-direction: column;
-    gap: 2.5rem;
-    color: var(--ink);
-    padding-bottom: 4rem;
-}
-
-.product-breadcrumb {
-    font-size: 0.9rem;
-    color: var(--muted);
-    display: flex;
-    gap: 0.5rem;
-    align-items: center;
-}
-
-    .product-breadcrumb .current {
-        color: var(--ink);
-        font-weight: 600;
-    }
-
-.product-main {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
-    gap: 2.5rem;
-    align-items: start;
-}
-
-.product-gallery {
-    display: flex;
-    flex-direction: column;
-    gap: 1.5rem;
-}
-
-.main-image-card {
-    background: var(--panel);
-    border-radius: 24px;
-    padding: 2.5rem 2rem 3.5rem;
-    text-align: center;
-    position: relative;
-}
-
-    .main-image-card img {
-        width: 100%;
-        max-height: 360px;
-        object-fit: contain;
-    }
-
-.ghost-action {
-    position: absolute;
-    bottom: 1.5rem;
-    left: 50%;
-    transform: translateX(-50%);
-    border: 1px solid var(--border);
-    background: #fff;
-    padding: 0.4rem 1rem;
-    border-radius: 999px;
-    font-size: 0.85rem;
-    color: var(--ink);
-}
-
-.thumbnail-row {
-    display: grid;
-    grid-template-columns: repeat(4, minmax(0, 1fr));
-    gap: 1rem;
-}
-
-.thumbnail {
-    background: var(--panel);
-    border: 2px solid transparent;
-    border-radius: 16px;
-    padding: 0.75rem;
-    transition: border 0.2s ease;
-}
-
-    .thumbnail img {
-        width: 100%;
-        height: 70px;
-        object-fit: contain;
-    }
-
-    .thumbnail.is-active {
-        border-color: var(--accent);
-    }
-
-.product-details {
-    display: flex;
-    flex-direction: column;
-    gap: 1.4rem;
-}
-
-.product-meta {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.chip {
-    background: var(--soft);
-    color: var(--muted);
-    padding: 0.4rem 0.8rem;
-    border-radius: 999px;
-    font-size: 0.75rem;
-    font-weight: 600;
-}
-
-.icon-button {
-    border: 1px solid var(--border);
-    background: #fff;
-    border-radius: 50%;
-    width: 36px;
-    height: 36px;
-}
-
-.rating {
-    display: flex;
-    gap: 0.75rem;
-    align-items: center;
-    color: var(--muted);
-    font-size: 0.9rem;
-}
-
-.stars {
-    color: #f5b301;
-    font-size: 1rem;
-}
-
-.price {
-    font-size: 1.5rem;
-    font-weight: 700;
-}
-
-.option-group {
-    display: flex;
-    flex-direction: column;
-    gap: 0.6rem;
-}
-
-.option-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.option-label {
-    font-weight: 600;
-}
-
-.size-guide {
-    font-size: 0.8rem;
-    color: var(--accent-dark);
-    text-decoration: none;
-}
-
-.color-options {
-    display: flex;
-    gap: 0.6rem;
-}
-
-.color-option {
-    width: 24px;
-    height: 24px;
-    border-radius: 50%;
-    border: 2px solid var(--border);
-    background: #fff;
-}
-
-    .color-option.is-selected {
-        border-color: var(--accent-dark);
-        box-shadow: 0 0 0 3px rgba(199, 240, 77, 0.4);
-    }
-
-.color-olive {
-    background: #7aa03a;
-}
-.color-black
-
-{
-    background: #111;
-}
-.size-options
-{
-    display: flex;
-    flex-wrap: wrap;
-    gap: 0.6rem;
-}
-
-.size-option {
-    padding: 0.35rem 0.8rem;
-    border: 1px solid var(--border);
-    background: #fff;
-    border-radius: 8px;
-    font-size: 0.85rem;
-}
-
-    .size-option.is-selected {
-        background: var(--accent);
-        border-color: var(--accent);
-        font-weight: 600;
-    }
-
-.quantity-control {
-    display: inline-flex;
-    align-items: center;
-    border: 1px solid var(--border);
-    border-radius: 10px;
-    overflow: hidden;
-    width: fit-content;
-}
-
-.qty-btn {
-    width: 32px;
-    border: none;
-    background: #fff;
-    font-weight: 600;
-}
-
-
-    .quantity-control input
-
-{
-    width: 50px;
-    margin-right: 5px;
-    cursor: pointer;
-    border: none;
-    text-align: center;
-}
-
-.actions {
-    display: flex;
-    gap: 1rem;
-}
-
-.primary-btn,
-.secondary-btn {
-    padding: 0.65rem 1.5rem;
-    border-radius: 10px;
-    border: none;
-    font-weight: 600;
-}
-
-.color-box {
-    display: inline-block;
-    width: 20px;
-    height: 20px;
-    border: 1px solid #ccc;
-    margin-right: 5px;
-    .primary-btn
-
-{
-    background: #111827;
-    color: #fff;
-}
-
-.size-btn {
-    margin: 0 3px;
-    .secondary-btn
-
-{
-    background: var(--accent);
-    color: #1f2937;
-}
-
-.tabs {
-    background: #fff;
-    border-radius: 18px;
-    padding: 1.5rem;
-    border: 1px solid var(--border);
-}
-
-.tab-menu {
-    display: flex;
-    gap: 1.5rem;
-    border-bottom: 1px solid var(--border);
-    padding-bottom: 0.75rem;
-}
-
-.tab-link {
-    border: none;
-    background: transparent;
-    padding-bottom: 0.75rem;
-    font-weight: 600;
-    color: var(--muted);
-}
-
-    .tab-link.is-active {
-        color: var(--ink);
-        border-bottom: 3px solid var(--accent);
-    }
-
-.tab-content {
-    padding-top: 1.5rem;
-    color: var(--muted);
-    line-height: 1.6;
-}
-
-.tab-panel {
-    display: none;
-}
-
-    .tab-panel.is-active {
-        display: block;
-    }
-
-.section-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-    .section-header a {
-        font-size: 0.85rem;
-        color: var(--accent-dark);
-        text-decoration: none;
-    }
-
-.card-row {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(210px, 1fr));
-    gap: 1.5rem;
-    margin-top: 1.5rem;
-}
-
-.product-card {
-    background: var(--card);
-    border: 1px solid var(--border);
-    border-radius: 16px;
-    overflow: hidden;
-    box-shadow: 0 8px 20px rgba(15, 23, 42, 0.06);
-}
-
-    .product-card img {
-        width: 100%;
-        height: 180px;
-        object-fit: cover;
-    }
-
-    .card-body
-{
-    padding: 1rem;
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-}
-
-.badge {
-    background: #fef3c7;
-    color: #92400e;
-    border-radius: 999px;
-    font-size: 0.7rem;
-    padding: 0.25rem 0.6rem;
-    width: fit-content;
-}
-
-.card-price {
-    font-weight: 700;
-}
-
-.mini-btn {
-    background: var(--accent);
-    border: none;
-    padding: 0.4rem 0.8rem;
-    border-radius: 8px;
-    font-weight: 600;
-    font-size: 0.8rem;
-}
-
-.feature-row {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
-    gap: 1.5rem;
-    background: #f8fafc;
-    padding: 1.5rem;
-    border-radius: 16px;
-}
-
-.feature
-{
-    display: flex;
-    gap: 1rem;
-    overflow-x: auto;
-    gap: 0.75rem;
-    align-items: center;
-}
-    .feature-icon
-{
-    background: var(--accent);
-    width: 36px;
-    height: 36px;
-    border-radius: 10px;
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.newsletter {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    gap: 2rem;
-    background: #0f172a;
-    color: #fff;
-    padding: 2.5rem;
-    border-radius: 20px;
-}
-
-    .newsletter .primary-btn {
-        background: var(--accent);
-        color: #0f172a;
-    }
-
-.newsletter-graphic {
-    font-size: 4rem;
-    color: var(--accent);
-}
-
-.top-picks {
-    display: flex;
-    flex-direction: column;
-    gap: 1.5rem;
-}
-
-@media (max-width: 768px) {
-    .product-main {
-        grid-template-columns: 1fr;
-    }
-
-    .newsletter {
-        flex-direction: column;
-        align-items: flex-start;
-    }
-}
